fix: refactor to ensure cross compile
