+++ internal/core/infra/platform/darwin/monitor_select_overlay_darwin.m
